+++ STAY/code/common/components.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "graphics/renderSystem.hpp" // IWYU pragma: export
+#include "physics/rigidBody.hpp" // IWYU pragma: export
+#include "physics/joint.hpp" // IWYU pragma: export
+#include "physics/collider.hpp" // IWYU pragma: export
