+++ STAY/code/game/component/list.hpp
@@ -1,10 +1,7 @@
 #pragma once
 
-#include "render.hpp" // IWYU pragma: export
-#include "../../common/physics/collider.hpp" // IWYU pragma: export
+#include "components.hpp" // IWYU pragma: export
 #include "player.hpp" // IWYU pragma: export
-#include "../../common/physics/rigidBody.hpp" // IWYU pragma: export
 #include "hook.hpp" // IWYU pragma: export
-#include "../../common/physics/joint.hpp" // IWYU pragma: export
 #include "playerDebug.hpp" // IWYU pragma: export
 #include "dash.hpp" // IWYU pragma: export
