--- STAY/code/game/system/list.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "OrderedRenderSystem.hpp" // IWYU pragma: export
-#include "PhysicsDebugSystem.hpp" // IWYU pragma: export
-#include "PhysicsSystem.hpp" // IWYU pragma: export
-#include "playerSystem.hpp" // IWYU pragma: export
-#include "hookSystem.hpp" // IWYU pragma: export
-#include "debugSystem.hpp" // IWYU pragma: export
-#include "dashSystem.hpp" // IWYU pragma: export
